very basic framework for a request page
been p busy today so haven't had much time, but this is a incredibly
basic layout for a requests page. the only current issue is that the
document.getElementById() calls in sendRequesnt end up null. I aim
to fix this either later today or tomorrow, but if anyone else wants to
take a crack at it go ahead.